+++ CHANGES.md
@@ -13,6 +13,8 @@
 
 - [CHANGE] `boost::optional` を利用していた部分を全て `std::optional` に変更
   - @melpon
+- [CHANGE] SRTP keying material を取得する機能を削除
+  - @melpon
 - [UPDATE] VPL を v2.13.0 に上げる
   - @voluntas
 - [UPDATE] CMake を 3.30.5 にあげる